+++ src/main/java/com/eliteessentials/commands/hytale/HytaleReplyCommand.java
@@ -4,6 +4,7 @@ import com.eliteessentials.config.ConfigManager;
 import com.eliteessentials.EliteEssentials;
 import com.eliteessentials.permissions.Permissions;
 import com.eliteessentials.services.MessageService;
+import com.eliteessentials.services.SpyService;
 import com.eliteessentials.util.CommandPermissionUtil;
 import com.eliteessentials.util.MessageFormatter;
 import com.hypixel.hytale.component.Ref;
@@ -125,6 +126,12 @@ public class HytaleReplyCommand extends AbstractPlayerCommand {
         // Broadcast to console if enabled
         if (configManager.getConfig().msg.broadcastToConsole) {
             logger.info("[MSG] " + senderName + " -> " + targetName + ": " + message);
+        }
+
+        // Notify DM spies
+        SpyService spyService = EliteEssentials.getInstance().getSpyService();
+        if (spyService != null) {
+            spyService.notifyDmSpy(sender.getUuid(), target.getUuid(), senderName, targetName, message);
         }
     }